+++ OpenInvoicePeru/OpenInvoicePeru.Comun/Constantes/Respuestas.cs
@@ -4,6 +4,7 @@
     {
         public const string FormatoFecha = "yyyy-MM-dd";
         public const string FormatoNumerico = "###0.#0";
+        public const string EncodingIso = "ISO-8859-1";
     }
 }
