+++ OpenInvoicePeru/OpenInvoicePeru.Firmado/ICertificador.cs
@@ -1,0 +1,9 @@
+﻿using OpenInvoicePeru.Comun.Dto.Intercambio;
+
+namespace OpenInvoicePeru.Firmado
+{
+    public interface ICertificador
+    {
+        FirmadoResponse FirmarXml(FirmadoRequest request);
+    }
+}
