+++ OpenInvoicePeru/OpenInvoicePeru.Firmado/ISerializador.cs
@@ -1,0 +1,8 @@
+﻿namespace OpenInvoicePeru.Firmado
+{
+    public interface ISerializador
+    {
+        string GenerarXml<T>(T objectToSerialize);
+        string GenerarZip(string tramaXml, string nombreArchivo);
+    }
+}
