+++ OpenInvoicePeru/OpenInvoicePeru.Firmado/OpenInvoicePeru.Firmado.csproj
@@ -51,6 +51,9 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Certificador.cs" />
+    <Compile Include="ICertificador.cs" />
+    <Compile Include="ISerializador.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Serializador.cs" />
   </ItemGroup>
